+++ app/controllers/api/v1/user_schedule_controller.rb
@@ -1,30 +1,8 @@
 class Api::V1::UserScheduleController < ApplicationController
-    before_action :userSignedin?, only: [:create,:update,:destroy] #セッションの確認
+  before_action :userSignedin?, only: [:destroy] #セッションの確認
+  before_action :calendarOwn?, only: [:destroy] #カレンダーの所有者か確認
     
   def index
-      schedules = Schedule.all.order('id DESC').searchAPI("position",params[:p]).searchAPI("fr",params[:q]).first(50)
-      result = [];
-      schedules.each do |pas|
-          result.push(
-                      id:pas.id,
-                      title:pas.title,
-                      CoNum:pas.CoNum,
-                      teacher:pas.teacher,
-                      semester:pas.semester,
-                      position:pas.position,
-                      grade:pas.grade,
-                      status:pas.status
-                      )
-      end
-      
-      
-      render json: JSON.pretty_generate({
-                                        status:'SUCESS',
-                                        api_version: 'v1',
-                                        data_count:result.length,
-                                        schedules: result
-                                        
-                                        })
                 
   end
 
@@ -38,7 +16,8 @@ class Api::V1::UserScheduleController < ApplicationController
   end
   
     def destroy
-        ins = UserScheduleRelation.find_by(user_id:@user.id,schedule_id:params[:schedule_id],reges_grade:params[:grade])
+      if @userSession && @calendarOwn
+        ins = CalendarScheduleRelation.find_by(calendar_id:params[:calendarId],schedule_id:params[:schedule_id],reges_grade:params[:grade])
         if ins.destroy
             render json: JSON.pretty_generate({
                                               status:'SUCCESS',
@@ -53,6 +32,6 @@ class Api::V1::UserScheduleController < ApplicationController
             
                                                 })
         end
-      
+      end
   end
 end
